// enums/vuetify.ts
export enum BtnsVariants {
  Elevated = 'elevated',
  Outlined = 'outlined',
  Flat = 'flat',
  Text = 'text',
  Tonal = 'tonal',
  Plain = 'plain',
}

export enum Colors {
  Info = 'info',
  Warning = 'warning',
  Secondary = 'secondary',
  Primary = 'primary',
  Background = 'background',
  Easy = 'easy',
}

export enum Sizes {
  XSmall = 'x-small',
  Small = 'small',
  Default = 'default',
  Large = 'large',
  XLarge = 'x-large',
}

export enum Locations {
  Start = 'start',
  End = 'end',
  Left = 'left',
  Right = 'right',
}